backend/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
redis>=5.0.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Background Tasks
celery>=5.3.6

# File Handling
pillow>=10.2.0
python-magic>=0.4.27

# HTTP Client
httpx>=0.27.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# WebSocket support
websockets>=12.0

# Email validation
email-validator>=2.1.0
